move page title resolver to main folder
