+++ projects/core/src/page/title/facade/index.ts
@@ -1,5 +1,5 @@
 export * from './page-title.service';
-export * from './resolvers/page-title.resolver';
+export * from './page-title.resolver';
 export * from './resolvers/product-page-title.resolver';
 export * from './resolvers/category-page-title.resolver';
 export * from './resolvers/content-page-title.resolver';
