+++ projects/core/src/page/title/facade/resolvers/category-page-title.resolver.ts
@@ -9,7 +9,7 @@ import {
   PageType,
   PaginationModel
 } from '../../../../occ/occ-models/occ.models';
-import { PageTitleResolver } from './page-title.resolver';
+import { PageTitleResolver } from '../page-title.resolver';
 
 @Injectable({
   providedIn: 'root'
